Remove RuleManager and RuleManagerImpl #41 https://github.com/marinmitev/smarthome/issues/41
